Noah D first 6 tasks commit
Might break stuff. Get ready to undo this commit!
* Initial update

Includes:
-Stamina bar and system
-Frustration system with face indicator and slider
-Misc related stuff like a killzone that reduces stamina and has a spawn point nearby, some basic checkpoints that set stamina to max and respawn player there at 0 stamina, and a few other things

Things to note: I've replaced the killzones in the main scene with the new ones, and the new prefab can be found in Assets/Prefab/Script-Heavy Objects
    Because of a bug where RobotBoy would count EVERY 2D collider as ground, ground is now defined by a single layer (ground) instead of everything, as it was. Ground must now be put on the correct layer. Done for the current level but must be done in future by level designers. Discord with any questions
    Because of the way I monitor player movement and stamina drain, I needed a unique version of the character controlling scripts. Because of that and because of a few other things I also have a unique CharacterRobotBoy prefab in the Assets/Prefab folder. Make sure you use that if you create a new scene
   The GUI is also in the prefab folder. Drag it onto a new scene to add it, and given the presence of the custom version of the player character it should just work. I've already added it to the main scene, obviously
   GUI can be modified, however some of the objects know locations in space where they're supposed to be and where they're supposed to go. If designers/artists need to modify the positions/sizes of the GUI elements they'll need to talk to me so they know how to make sure everything still works once they're done.
   Message me on Discord with any questions.

This update contains everything asked for the week 1 and 2 task

* Damage collider script + Ease of use + tune up

Made a few changes to make the features a bit easier to use to those without any programming experience, most things should be tweaked from the inspector view in unity
A notable exception is the GameConst script, which must be edited in a text editor because of the way it works (doesn't extend monobehaviour)

* Fixed some stuff that merging with the master kinda broke

* Working Dialogue System Initial Commit

Has and implements a working dialogue system inside of testScene.
Features:
-Dialogue System
-Prototype quest:Wizard apple quest
-Minor fixes, for example anchoring simple UI elements correctly

* Inevitable bug fixes

Fixed the GUI on the main scene and added the correct camera to the testing scene that contains the test quest.

* Swamp level task commit

-Ladder
-Quicksand
-Basic UI visuals

* Revert "Merge branch 'master' into NoahDProgrammingCorner"

This reverts commit c2b3ad2481d739cb29fa4580ca50b05c2586bf3b, reversing
changes made to 5509e06fa5ec1b277821318706e219fec89285d6.

* Rogue quest intitial commit

Has the beginnings of the rogue quest. Would continue but my task for next week is to get the playercontroller up and running with new animations