Re-Added Player & Animations
Player is added. Glitches in the Animation are present at the moment so please ignore them. I'll have to re-do the animations as well as re-build the Character possibily in Spline or re-do Anima 2D because of porting glitches from 3.1 to 2.4. I'll re-focus on the player once Everything else is set.